solr-feeder/requirements/prod.txt
Flask
Flask-Restplus
python-dotenv
requests
flask_sqlalchemy
psycopg2-binary
